First shot for les maisons phenix.
Although it might change a lot, the idea was to setup quickly an app
for Myriam and Sandra to upload their material.

It will allow us (Alex + Sarah) to work with actual content.

I took this opportunity to test django celery so we can transcode on the
fly the sound files (yes, we **need** to work with free codecs).

The root view at <http://localhost:8000/> re-uses the idea of
lazy-loading that was developed for active archives: Resources are not
immediatly loaded but links (`<a rel="embed" />`) are spidered with
a javascript that launch a celery task. The task returns the resource
info, and possibly transcode the resources if not done yet. Using celery
shipped views the javascript gets back a json with the status of the
task, and if ready, injects the actual HTML.

In this case, we start from non web-friendly audio or video resources,
convert them to ogg, and inject `<audio>` or `<video>` elements in the
page.