Switched to a more flexible model inspired by [active archives](http://activearchives.org) experiments for annotating the sound files.
The idea is to do some kind of "media" wiki, not in the sense of
mediawiki but in the sense of a wiki on media; only sound at the moment.

Very stupid: it is just a django model with a filefield to upload sound,
and a textfield to annotate the sound using markdown and some custom
markdown plugin to create semantic markup. When the instance is saved,
the HTML+RDFa is parsed into the RDF store. That way we can avoid
tabular interfaces and let us take arbitrary notes. The HTML **is** the
database.

I was trying to put reuse the pieces we've been developping for active
archives, when I figured out DRY is a good thing, except that the D here
stands for "Do", as in "Do Repeat yourself".

Let see how it goes...